Rename module `util` -> `valvis`.
It turned out to not be a pile of utility classes.
