+++ src/collections/package.json
@@ -16,6 +16,6 @@
 
   "dependencies": {
     "@this/typey": "*",
-    "@this/util": "*"
+    "@this/valvis": "*"
   }
 }
