+++ src/net-util/package.json
@@ -20,7 +20,7 @@
     "@this/loggy-intf": "*",
     "@this/quant": "*",
     "@this/typey": "*",
-    "@this/util": "*",
+    "@this/valvis": "*",
     "mime": "^4.0.1",
     "range-parser": "^1.2.1",
     "selfsigned": "^2.4.1",
