+++ src/quant/package.json
@@ -15,6 +15,6 @@
 
   "dependencies": {
     "@this/typey": "*",
-    "@this/util": "*"
+    "@this/valvis": "*"
   }
 }
