--- src/util/README.md
@@ -1,10 +1,0 @@
-@this/util
-==========
-
-Low-dependency utility classes that have no better place to live.
-
-- - - - - - - - - -
-```
-Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
-SPDX-License-Identifier: Apache-2.0
-```
