+++ src/valvis/README.md
@@ -1,0 +1,11 @@
+@this/valvis
+============
+
+Generic visitor mechanism for JavaScript object graphs, along with related
+helper classes and utility functionality.
+
+- - - - - - - - - -
+```
+Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
+SPDX-License-Identifier: Apache-2.0
+```
