+++ src/valvis/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/util",
+  "name": "@this/valvis",
   "version": "0.0.1",
   "type": "module",
   "private": true,
